SberbankTest1/requirements.txt
allure-pytest==2.5.0
allure-python-commons==2.5.0
atomicwrites==1.1.5
attrs==18.1.0
click==6.7
colorama==0.3.9
enum34==1.1.6
FoxPuppet==1.0.1
lxml==4.2.2
more-itertools==4.2.0
namedlist==1.7
pluggy==0.6.0
precisely==0.1.5
py==1.5.4
pytest==3.6.3
pytest-click==0.1
pytest-firefox==0.1.1
selenium==3.13.0
six==1.11.0
